docs: added references to scipy
