+++ docs/source/modules/transforms/transforms.rst
@@ -1,0 +1,11 @@
+=====================
+floodlight.transforms
+=====================
+
+Some useful transformers
+
+.. toctree::
+   :maxdepth: 1
+   :caption: Transforms
+
+   filter
